# bpo-40137: Micro-optimize _PyType_GetModuleByDef()