--- a/Misc/NEWS.d/next/Core and Builtins/2021-04-13-22-48-56.bpo-40137.Z77R2P.rst
+++ b/Misc/NEWS.d/next/Core and Builtins/2021-04-13-22-48-56.bpo-40137.Z77R2P.rst
@@ -1,0 +1,2 @@
+Micro-optimize _PyType_GetModuleByDef() to make it 2.3 ns faster (43.2 ns +-
+0.7 ns -> 40.9 ns +- 1.0 ns). Patch by Victor Stinner.