Make _PyType_GetModuleByDef() 2.3 ns faster.

Benchmark: 43.2 ns +- 0.7 ns -> 40.9 ns +- 1.0 ns: 1.05x faster

./python -m pyperf timeit \
    --duplicate=4096
    -s "from functools import lru_cache; f = lru_cache(lambda: 42)" \
    "f()"

Changes:

* _PyType_GetModuleByDef(): add fast-path for tp_mro[0] and use
  _PyType_HasFeature().
* Add a new pycore_moduleobject.h internal C API header file.
* Add _PyModule_GetDef() and _PyModule_GetState() which can be
  inlined without LTO and don't check invalid argument at runtime.
* Replace PyModule_GetState() with _PyModule_GetState() in _abc,
  _array, _operator, _pickle, _queue, _random, _struct and os
  extension modules. The _array, _queue and _struct extensions are
  now built with Py_BUILD_CORE_MODULE macro defined.

<!--
Thanks for your contribution!
Please read this comment in its entirety. It's quite important.

# Pull Request title

It should be in the following format:

```
bpo-NNNN: Summary of the changes made
```

Where: bpo-NNNN refers to the issue number in the https://bugs.python.org.

Most PRs will require an issue number. Trivial changes, like fixing a typo, do not need an issue.

# Backport Pull Request title

If this is a backport PR (PR made against branches other than `master`),
please ensure that the PR title is in the following format:

```
[X.Y] <title from the original PR> (GH-NNNN)
```

Where: [X.Y] is the branch name, e.g. [3.6].

GH-NNNN refers to the PR number from `master`.

-->


<!-- issue-number: [bpo-40137](https://bugs.python.org/issue40137) -->
https://bugs.python.org/issue40137
<!-- /issue-number -->