@vstinner 2021-04-13
> Replace PyModule_GetState() with _PyModule_GetState() in (...) extension modules

Honestly, the runtime check on PyModule_GetState() argument could replaced with an assertion. But I prefer to not change in as part of this PR. See https://bugs.python.org/issue37406 for my previous attempt: "Disable runtime checks in release mode".

But _PyModule_GetState() remains useful to ensure that the code can be inlined, especially on macOS which doesn't use LTO: https://bugs.python.org/issue41181