@rhettinger 2021-04-21
This looks good to my eyes but I haven't deeply thought about what the slow path is doing and what cases it handling.  So, this is a 95% sign-off.