@vstinner 2021-04-21
Hum. I'm not 100% sure about the assumptions that I made in my optimization. I plan to write first preparation changes to make the assumptions reliable :-)