@vstinner 2021-04-21
Also, PR #25492 makes this optimization less important.