@vstinner 2021-04-21
I splitted this PR into 3 smaller PRs with a few minor fixes: https://bugs.python.org/issue40137#msg391547

The main change is that I added a runtime check to ensure that a type MRO cannot be empty. I close this PR.